requisitos restantes
