+++ src/context/Provider.jsx
@@ -15,6 +15,19 @@ const RecipeProvider = ({ children }) => {
   const [places, setPlaces] = useState([]);
   const [recipesByPlace, setRecipesByPlace] = useState([]);
   const [area, setArea] = useState();
+  const [ingredientsList, setIngredientsList] = useState([]);
+  const [valueButton, setValueButton] = useState('');
+  const [dataFood, setDataFood] = useState([]);
+  const [filterFood, setfilterFood] = useState([]);
+  const [dataCategoriesFood, setDataCategoriesFood] = useState([]);
+  const [filterCategoriesFood, setFilterCategoriesFood] = useState([]);
+  const [dataCategories, setDataCategories] = useState([]);
+  const [filterDataCategories, setFilterDataCategories] = useState([]);
+  const [toogle, setToogle] = useState(false);
+  const [allValue, setAllValue] = useState('');
+  // const [selectedIngredient, setSelectedIngredient] = useState('');
+  const [previousIsExploreIngredients, setPreviousIsExploreIngredients] = useState(false);
+  const [filterDrinksIngredients, setFilterDrinksIngredients] = useState([]);
 
   useEffect(() => {
     async function setDataRecipes() {
@@ -99,6 +112,32 @@ const RecipeProvider = ({ children }) => {
     setRecipesByPlace,
     area,
     setArea,
+    ingredientsList,
+    setIngredientsList,
+    // selectedIngredient,
+    // setSelectedIngredient,
+    previousIsExploreIngredients,
+    setPreviousIsExploreIngredients,
+    filterDrinksIngredients,
+    setFilterDrinksIngredients,
+    valueButton,
+    setValueButton,
+    dataFood,
+    setDataFood,
+    filterFood,
+    setfilterFood,
+    dataCategoriesFood,
+    setDataCategoriesFood,
+    filterCategoriesFood,
+    setFilterCategoriesFood,
+    dataCategories,
+    setDataCategories,
+    filterDataCategories,
+    setFilterDataCategories,
+    toogle,
+    setToogle,
+    allValue,
+    setAllValue,
   };
   return (
     <RecipeContext.Provider value={ objContext }>
